Refactor Parser Interface
- part of  #761
- the SymbolicExpression class is newly introduced,
which internally manages the parser. It can evaluate
the symbolic expression, as well as form the first and
second derivatives.
- Required variables and constants for the evaluation
of the symbolic expression must be passed as arguments
to the corresponding Evaluate function. (Constants and
variables can thus no longer be set by the parser member
functions SetValue and AddVariable in the course of the
code as before).
- To enable copying of the symbolic expression, a
Copy Constructor, Copy Assignment Operator, Move Constructor,
Move Assignment Operator and Destructor are added
if required. (Rule of 5)
- Inheritance between Lexer and Parser is removed.
(Composition via inheritance)
- In an input file of a 2D test case "z" was removed
as third dimension. Results did not need to be adjusted
because this value was previously set to zero internally.