Adding LINALG::SOLVER interface to FSI
This implements a solver interface to LINALG, by adding
additional input parameters and the related solver calls
for the fsi module.

One test has been adjusted to use the new interface. As
the underlying Epetra_Operator is a block matrix, only
a direct solver is so far used. The algebraic multigrid
preconditioner for FSI problems is in the making.

The new interface should be tested with more than one
test. Just added the interface in under 1 hour, which
went surprisingly smooth (just a piece of cake ;) ).

Related to #870
Part of #887