Implement EvaluateNeumann in new solid elements
This commit implements the evaluation of Neumann loads for the new solid
element implementation. The function is implemented as a free function
and can thus be used also in other parts of the code.

This commit further introduces a new free function that iterates within
an element over each Gauss points and evaluates the jacobian mapping.

Issue #868