Introduce recursive coordinate bisection rebalance
Adding an algorithm for partitioning/rebalancing based on a
graph and it's coordinates. This is useful for a partitioning, where
certain things with spatial distance need to be on the same
processor, e.g. beam-solid interaction. The respective rebalancing
methods are enhanced to be able to also include coordinates of a
discretization and make use of it.

Closes !1159, #808.