Simplify usage of CSV Runtime Writer
All setup tasks are moved to the constructor, such that less side
effects of the various methods occur.

Additionally, an inheritance structure is introduced in an
implementation class, that is accessed via an interface class. The
derived classes distiguish between proc 0 that does all writing and
other procs that do nothing