FSI: move some xffsi tests to NOX implementation
The NOX implementation is the main and should be the only one
implementation for monolithic FSI. This commit moves as many tests from
the separate XFFSI implementation without NOX over to the NOX-based
implementation.

Remaining tests require more work, as one needs to check if the xfem
DOFs have changed if running in "FullNewton" mode.