Introduce probdim to scatra ele boundary calc
Like in scatra ele calc, the dimension of the problem is introduced.
Furthermore, the number of space dimension (nsd) is split into the
number of space dimension of the element and of the entire problem. This
allows to have e.g. 1D elements in a 3D space. So far, the number of
space dimensions was assumed to be one order higher as the element.
However, in the test 'scatra_3D_quad4_hex8_s2i_constperm.dat' this was
not the case, s.t. the metric tensor was wrongly computed. This is fixed
now.