ScaTra Ele Boundary: Computation of normal vector
Rework the computation of the normal vector of a boundary element:
- Allow the computation of an 1D boundary element in a 3D space
- Throw an error, if the computation is requested for NURBS elements
(was marked 'ToDo: this is only a temporary solution in order to have
something here.')

Remove post processing of boundary flux in 'scatra_nurbs9_channel.dat',
'elch_nurbs9_linear.dat', and 'elch_nurbs9_tafel.dat'.
This would lead to a dserror now.