Move init functionality to the constructor
Some cleanup of scatra methods where functionality that was performed
to setup the classes is moved from the init method to the constructor.

This also enables to set the number of dofset for an added field
directly after adding it to the corresponding discretization and
therefore eases the dofset handling.