Fix bugs in vtk output
which leads to invalid vtu output

1. Bug:
The cell_connectivity_ vector is never cleared such that it is never
rebuilt when new nodes are added (e.g. through redistribution) This
leads in Paraview to errors like: `Cannot read cell connectivity from
Cells in piece 0 because the "connectivity" array is not long enough.`
Fix: clear the cell_connectivity_ vector in
SetGeometryFromDiscretization which is called when the parallel
distribution changed

2. Bug:
The ConsistencyCheck() is insufficient. It checks the cell_connectivity_
against the number of cells. But the connectivity must also be as long
as the largest referenced value in the cell offset.