Extract element formulation in solid elements
With this new structure, we extract the element formulation from the
solid element evaluation. The solid element calculation class is now
templated for the element formulation. This allows us to reduce a lot
duplicate code. New element technologies (like F-Bar) can now be written
in a few lines of code.