Initialize norm of rhs in baci_fsi_xfem_monolithic
The value was used uninitialized and causing a jump depending on
an uninitialized value. This initializes the value with 0.0 such that
this conditional jump is now (at least) defined.

Note: This issue might point to a deeper problem in the algorithm that
is not fixed with this MR. This just fixes this undefined behavior.

Issue #842