CI: Enable caching of previous pipeline builds
Previously, we always deleted the build folder in order to get a clean
build for every pipeline. However, we can make use of the previous
build as make/ninja will make sure that any real changes will be
rebuilt so that we get the same result as from a clean built but
potentially faster. We only need to clean the cmake cache, which is
done with the --fresh option, as the cmake cache will keep variables
in the cache even though they have been removed from the CMakeLists.

In the rare case that an executable is not built anymore, it can happen
that scripts that have not been updated accordingly will still use the
executable present in the build folder. To catch this case, a clean
build is done in the nightly pipeline.

See #1228.