LazyPtr class to delay construction
This class allows to specify how an object is constructed without
performing the construction right away. This design allows to state
that object A needs object B for construction, before object B exists.
If both A and B are wrapped in LayzPtr, then construction of A will
trigger construction of B automatically.