Remove parallel redistribution from monolithic FSI
The overlapping domain decomposition in monolithic FSI had to be
redistributed to design a "hybrid preconditioner", c.f. PhD thesis by
Matthias Mayr. This feature has been removed, so we don't need the code
for redistribution anymore. This commit removes hopefyll all code
related to the construction of a monolithic graph of the FSI problem as
well as the redistribution.

Part of #1289