Remove definition of nurbs27 from
GetVtkCellTypeFromFourCElementShapeType as nurbs should not be queried
from this method
Throw an error if it is called by an element with different cell type.
Rename the method according to the new naming conventions.
Remove the results of node-based output from one test case using
nurbs27 discretization as it obviously was wrong.