Preserve test goal despite change in solver
Changing the preconditioner in an FSI simulation can slightly change the
results. This is particularly pronounced in this test case, which is
designed to fail convergence of the nonlinear solver in order to
exercise the rescue mechanism of halving the time step size in case of
non-convergence of the Newton scheme.

To preserve the purpose of the present test, the number of time steps is
reduced, such that the final time step is repeated once due to an
unconverged nonlinear solver. This is important, since the number of
repetitions is tested (and should be different than the ideal case of
no repetition).