Introduce tensor interpolation routine
- Added interpolation scheme for invertible second-order tensors (3x3)
as a templated class within "core", as detailed in Satheesh et al. 2024
(10.1002/nme.7373). This algorithm preserves the structure of the
underlying tensors, along with relevant tensor characteristics.
- The corresponding template parameter specifies the dimension of the
tensor locations (e.g. interpolation over one dimension), and the
constructor allows for the specification of an interpolation
order(linear, quadratic,...).
- Implemented several other methods used during interpolation, such as
the polar decomposition of a 3x3 matrix, the computation of rotation
vectors from rotation matrices and vice-versa.
- Added unit tests for the interpolation between symmetric and
nonsymmetric tensors.