Further unify pack and unpack
The operations are now symmetric for most types. It is no longer
necessary to go through extra steps to extract data of ParObject into a
vector<char> before parsing this buffer. This is mainly achieved by
removing the SizeMarker. Debugging is aided in assert mode by writing
the hash code of the types added to the buffer and checking that
_exactly_ that data is extracted. Thus, we no longer have cases where we
need to e.g. cast bools to int before adding/extracting.