Add docker build for image with pre-built 4C
For now, the 4c-dependencies-ubuntu24.04 image is used as a base where
we pull and build 4C. This allows user to pull the docker image and
check out the 4C executable without needing to set up the build
environment and building 4C themselves. Both the source and build are
retained in the image. Later, one can reduce the image size by only
including the essentials for the 4C executable.

The docker build runs after the nightly workflow if it was successful.

Since we do not have a versioning yet, the image is tagged `latest`.