Add min ele option to initial rebalancing
This PR adds a user option to specify a minimum
amount of elements, which have to be on a processor
to be used for calculation.

This is helpful for problems with a dominating domain
consuming a lot of processors, while smaller ones might
be oversubscribed by the number of processors given.